use dynamic import without ssr to load simulator page
